EPO2_V5/api-flask/requirements.txt
Flask
flask-mysqldb
flask-cors
mysql-connector-python
requests